requirement.txt
asgiref==3.3.1
astroid==2.4.2
Django==3.1.5
django-ckeditor==6.0.0
django-js-asset==1.2.2
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.2
pylint-django==2.4.2
pylint-plugin-utils==0.6
python-decouple==3.4
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
